관리자 NONE 팀 관리 기능 수정

+++ Meeron/Meeron/ViewModel/TeamViewModels/TeamManagementViewModel.swift
@@ -10,7 +10,7 @@ import RxSwift
 
 class TeamManagementViewModel {
     var participants:[WorkspaceUser] = []
-    var nowTeam:Team
+    var nowTeam:Team?
     var newTeamName:String
     
     let participantsSubject = BehaviorSubject<[WorkspaceUser]>(value: [])
@@ -18,22 +18,35 @@ class TeamManagementViewModel {
     
     let teamRepository = TeamRepository()
     let disposeBag = DisposeBag()
-    init(participants:[WorkspaceUser], nowTeam:Team) {
+    
+    init(participants:[WorkspaceUser], nowTeam:Team?) {
         self.participants = participants
         self.nowTeam = nowTeam
-        self.newTeamName = nowTeam.teamName
+        self.newTeamName = nowTeam?.teamName ?? "NONE"
         participantsSubject.onNext(participants)
     }
     
     func loadParticipant() {
-        teamRepository.loadUsersInWorkspaceTeam(teamId: String(nowTeam.teamId))
-            .withUnretained(self)
-            .subscribe(onNext: { owner, users in
-                if let users = users {
-                    owner.participantsSubject.onNext(users.workspaceUsers)
-                    owner.participants = users.workspaceUsers
-                }
-            }).disposed(by: disposeBag)
+        if nowTeam == nil {
+            teamRepository.loadUsersWithoutTeam()
+                .withUnretained(self)
+                .subscribe(onNext: { owner, users in
+                    if let users = users {
+                        owner.participantsSubject.onNext(users.workspaceUsers)
+                        owner.participants = users.workspaceUsers
+                    }
+                }).disposed(by: disposeBag)
+        }else {
+            teamRepository.loadUsersInWorkspaceTeam(teamId: String(nowTeam!.teamId))
+                .withUnretained(self)
+                .subscribe(onNext: { owner, users in
+                    if let users = users {
+                        owner.participantsSubject.onNext(users.workspaceUsers)
+                        owner.participants = users.workspaceUsers
+                    }
+                }).disposed(by: disposeBag)
+        }
+        
     }
     
     func deleteProfile(data:WorkspaceUser) {
@@ -57,6 +70,10 @@ class TeamManagementViewModel {
     
     
     func deleteTeam(){
+        guard let nowTeam = nowTeam else {
+            return
+        }
+
         teamRepository.deleteTeam(teamId: nowTeam.teamId)
             .withUnretained(self)
             .subscribe(onNext: { owner, success in
@@ -66,6 +83,11 @@ class TeamManagementViewModel {
     }
     
     func patchNewTeamName() {
+        
+        guard let nowTeam = nowTeam else {
+            return
+        }
+        
         if newTeamName != nowTeam.teamName {
             teamRepository.patchNewTeamName(name: newTeamName, teamId: nowTeam.teamId)
         }
